ironpulse-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.1
pydantic[email]==2.7.4
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
boto3==1.34.131
celery[redis]==5.4.0
redis==5.0.7
python-multipart==0.0.9
Pillow==10.3.0
openai==1.35.7
resend==2.0.0
firebase-admin==6.5.0
